commited statistics add-ons

+++ TP1/Trabalho 1/transferFile.c
@@ -65,7 +65,7 @@ int sendFile(int fd)
 
         noBytes += read;
 
-        printf("Progress = %.2f\n", (noBytes/traF->fileSize * 100));
+        printf("Progress = %.2f\n", (float)(noBytes/traF->fileSize * 100));
         printf("seqNo sent = %d\n\n", (seqNo % 255));
 
         seqNo++;
@@ -111,7 +111,7 @@ int receiveFile(int fd)
 
         noBytes += read;
         printf("seqNo received = %d\n", (seqNo % 255));
-        printf("Progress = %.2f\n", (noBytes/fileSize * 100));
+        printf("Progress = %.2f\n", (float)(noBytes/fileSize * 100));
 
         if ((written = fwrite(buffer, sizeof(char), read, traF->file)) > 0) {
             printf("%d bytes written to file!\n\n", written);
